backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Pillow==8.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7
gunicorn==21.2.0 